Refactor quota slot accounting into control plane
